Move commandline handling to separate file
This structures the source code a little bit better, but the main reason is
that compilation time improves because the (very large) args.hpp header is
no longer included from main.cpp, only from cmdline.cpp. So as long as
that file does not change, the args.hpp won’t be read at all.