Force correct read length on test data
The read length is detected incorrectly at the moment. By using -r 300, we
force it to the correct value.

This changes mapping results. Some mapping qualities change from 0 to 60.
Also, the second 'rescuable.42' read can no longer be rescued for some
reason, so we tweak the sequence a bit (verified manually that it still
triggers rescue).