Do not store randstrobes with count==1 in flat vector
They are not used anymore because they are stored directly in the hash
table. This makes the flat vector much smaller than previously.

This should already save a lot of memory when using an index that has been
saved to disk, but when the index is generated, there are still some
intermediate data structures that take up lots of space so the savings are
not as large.