Warn when binary was compiled without optimizations
When switching between branches, it often happens that CMake changes the
CMAKE_BUILD_TYPE to Debug, which makes the resulting binary very slow. I
have not been able to figure out, yet, why this is the case. I guess that
somehow the CMakeCache.txt file is invalidated by changing the branches, and
that this leads CMake to ignore the configured build type.

I would like to avoid that anyone benchmarks a strobealign binary compiled
without optimizations, so until we’ve solved this, I’d like to print a fat
warning when the build type is Debug.