Fix occasional incorrect soft clipping
Some reads starting with N bases in our test dataset were not soft clipped
at the 5' end, and instead a CIGAR "=" operator was (incorrectly) used.
This was apparently due to a bug in the loop that creates the CIGAR string.

The problem was less obvious because it was obscured by the soft-clip
algorithm being a bit pessimistic. In many cases, it returned an empty
alignment although a better solution was possible. In that case, the regular
alignment algorithm was triggered, wich then did give correct soft clipping.

To fix this, this commit changes the soft clip detection algorithm. Instead
of detecting where the maximum of the partial sum of the scores is, it tries
to find a segment in the alignment that has maximal sum of scores. This
works better for more complicated situations where, for example, both ends
need to be soft clipped.