Inline align_segment into get_alignment to fix string extraction
When get_alignment() calls align_segment(), it passes the reference extended
by 50 bp on either side. align_segment() then re-extracts the inner part
without the extension in order to compute the Hamming distance for that part
only. This was not done correctly in all cases.

The fix here is to inline align_segment into get_alignment and to rearrange
the code so that the Hamming distance is computed first using the un-extended
reference. If that is not succesful, the reference is extended in order to
compute the full alignment.