Introduce "end bonus" to penalize soft clipping
If an alignment extends until the end of the query, the end bonus is added
to the alignment score. (This should have the same effect as a soft clipping
penalty.)

Since SSW seems not to support an end bonus or soft-clipping penalty, we
postprocess the alignment: If an (ungapped) extension until the end reaches
a higher score (incl. the end bonus), then use that instead.