Use poolstl to sort randstrobes in parallel
See https://github.com/alugowski/poolSTL

Sorting the randstrobes is currently a bottleneck in index generation as it
does not run in parallel. This is an attempt to parallelize it.

poolstl’s sort uses regular std::sort under the hood.
We currently use pdqsort_branchless, which is about twice as fast as
std::sort, so parallel sorting breaks even with pdqsort_branchless at about
2-3 threads. It gets faster with more threads, but not as much as one would
perhaps expect. Here are the sorting runtimes for CHM13:

- 31 s with pdqsort_branchless
- 59 s with std::sort
- 34 s with parallel sort, 2 threads
- 24 s with parallel sort, 4 threads
- 23 s with parallel sort, 8 threads

Another issue is that sorting is no longer in place, so memory usage goes
up by a couple of gigabytes, which is another reason for me not to make this
change.