Add -DTRACE=ON option to show list of NAMs for debugging
If `cmake` is run with `-DTRACE=ON`, then some `#ifdef`s are active that
make strobealign print the list of found NAMs for each query.

I find myself adding this type of output quite often for debugging purposes,
and think it would be nice to just have this part of the code, disabled by
default, but easily enabled.