Do not use 0 as second hash
(This ensures that unmappable.41 in the test data really is unmappable)

Randstrobes that have no downstream partner at least w_min syncmers away get
their second hash set to 0, which means in the case of multi-context seeds
that the primary/main hash is also zero because it is chosen as the smaller
of the two.

When this is done both for the reference and for queries, we get spurious
hits for all randstrobes towards the ends of queries, which get mapped to the
end of the reference.

Using the hash of the primary syncmer also as hash for the second syncmer
gets rid of the problem.

Is-new-baseline: yes