Rename Hit struct to Match
We use the term "hit" both for
- a query strobemer successfully looked up in the index
- a locus on the query paired with a locus on the reference (this type of
  hit is merged into NAMs)

I think choosing the name for the latter to be "match" makes sense because
the M in NAM stands for "match". That is, we can then say that we merge
multiple overlapping matches into non-overlapping, approximate matches (=NAMs).

The flow would then be: hit -> match -> NAM.