Do a 'nested' lookup of randstrobes in mcs mode
In the index, full hits are located at a subinterval of the interval at
which partial hits are located. It is therefore conceptually nicer to
make use of the fact that the start of the two intervals are usually very
close to each other or even identical instead of looking up the full and
partial start indices separately.

To Do:
- we currently look up the partial hit first and then "tighten" the interval
to find the main if possible (by searching forward for the appropriate
index), but most of the time, we don’t need to find the partial hit and it
would be nicer to just have the main hit directly and only then do a
backwards linear search to find the partial hit if necessary.


Do linear search for full hash starting from partial hash position

In theory, this saves us looking up the correct bucket, but it does not
appear to make a difference in practice.