Produce a syncmer if any minimum is at the required position
This changes the way in which ties are handled during syncmer computation,
that is, what happens when there are multiple s-mers in the k-mer that have
the same minimum hash value (most likely because the s-mers are identical).

Previously, a syncmer would be returned only if the s-mer at position t
happened to be the rightmost minimum. Now it will be returned if any of the
s-mers with minimum hash is at positon t.

This increases the number of produced syncmers.