Add option for setting the number of indexing threads
This adds the hidden command-line option `--ithreads`, which allows to set a
different number of threads for indexing than for mapping/alignment.

This is mainly useful for debugging, for example, when also `--trace` is
used: For `--trace`, one should use `--threads 1` because the output of
multiple threads will otherwise interleave and not be useful. However,
because indexing then also happens with just one thread, it can be a bit
annoying to have to wait for that to happen. `--ithreads` solves this.