Stop checking anchors when we chain to one on the same diagonal
If we find a new anchor that we are going to chain to the current one on the same diagonal, then we assume that this is the correct one and skip checking other anchors.

This speeds up chaining significantly at no loss in accuracy.

This makes chaining typically on par speedwise with NAMs and often faster, except for fruitfly SIM0-SIM4, where it is still a little bit slower.

Is-new-baseline: yes