Let chunk size be in terms of nucleotides, not number of reads
Re-interpret --chunk-size to mean the number of *nucleotides* in a chunk.
This way, there are fewer reads in a chunk if the reads are longer.

This avoids the problem that we do not use all input threads for small input
datasets with long reads (if the input had 50000 reads, we would use at most
5 threads, for example).

Experiments suggest that chunk size can be relatively small without
adding too much overhead. Even for 100 nt reads and a chunk size of 1 read,
mapping is just a couple of percent slower than with bigger chunk sizes.

Results change slightly because the random seed uses the chunk index,
which changes.

Is-new-baseline: yes