Change the meaning of the -l and -u command-line parameters
They are now used to directly set w_min and w_max instead of the previous
(confusing) interpretation.

The names w_min and w_max are still used in the code at the moment, mostly
in order to avoid doing too many things at the same time, but also because
w_min and w_max is a bit easier to read than l and u.

This will be confusing for existing scripts!