New optimized indexing parameters
These were informed by a parameter optimization on SIM0 and SIM4, but to
avoid overfitting, we did not simply choose the best setting for each read
length, but a setting that improves results, does not slow down strobealign
too much, and is close to what we have at the moment.

The settings do generally not improve results on SIM6, where we instead
found -k 16 -s 12 -l 2 -u 2 to work well (all read lengths).

Also, remove canonical read length 125 because it became too similar to 100
and 150.