Log max no. of chunks in the OutputBuffer
The OutputBuffer could theoretically use a lot of RAM if one of the threads
takes a long time to process a chunk while the other threads working on
later chunks are fast (the later chunks get queued up until the earlier
chunk is finished). I have not seen this in practice, but it’s worth
logging this number.