Fix a memory leak when -r is used
When option `-r` is *not* provided, we read in the first 500 reads to
estimate the average read length and then call `InputBuffer::rewind_reset()`
to "rewind" to the beginning of the file. `InputBuffer` uses
`RewindableFile`, which keeps all data read from it in memory until
`rewind()` is called. (A RewindableFile can be rewound once.)

When `-r` is provided, we do not estimate average read length, never call
`InputBuffer::rewind_reset()`, never call `RewindableFile::rewind()` and
therefore grow the input buffer indefinitely until the end of file has been
reached or we are out of memory.

The solution chosen here is to always estimate the length of the input reads
even when -r is provided. This is somewhat of a workaround as it would be
nicer to fix InputBuffer/RewindableFile, but it is actually helpful because
this allows the user to check whether what they think the read length should
be matches what it actually is.